Webscraping - Superpro/requirements.txt
playwright==1.58.0
httpx>=0.28.0
loguru>=0.7.0
python-dotenv>=1.0.0
rich>=14.0.0
tenacity>=9.0.0
